Consolidated interface and types
